feat(trigger): Basic foreground and background color
This commit add basic support for customizingf foreground and background
color.
Currently a random primary color will set to the selection when called
through controller.
Note that the logic is far behind a normal editor when selection is
collapsed. Need to still set the attribute when selection is collapsed,
weired that the official appflowy-editor seems does not support it they
just ignored when selection is collapsed.