feat(*): Support record the last used selection
In may situations we lost focus and then want to insert some data or
restore the cursor, for example show a dialog to ask user insert
picture. This commit add support to record the last used not-null
selection so that the selection is accessible even we lost focus
before.

Maybe we should use quill, maybe.