fix(toolbar): Add trick to fix font size menu when caller used
This commit added some tricky thing to fix the font size menu not
popping up when caller use toolbar inside a chat_buttom_container.

Adding an empty font size option is a workaround and MUST be refactored
in the future, where we have the font size menu in dialog or bottom sheet.