feat(image): export image size
This commit added image size parameters to the image widget builder
function type, allowed to set and regard image width when rendering.

But, for now, the image's size doesn't change because editing the image
size will not trigger a refresh on the image widget UI. This should be
fixed in a near future.