fix(image): let image resizing with given size
Now image will resizing its appearance when user changed the bbcode
size.

Note that to provide a consist image width height ratio, more logic
should be set on limiting image size, currently this feature is
implemented by the editor user (developer) so that this package can
keep clean and lightwight on it.