feat(spoiler): toolbar button move to dialog
Seems the editor sometimes does not update cursor position correctly if
the tap hits on inner tappable child (e.g. an icon button) which cause
an embed not found exception when trying to save text to the editor.

This commit moved all toolbar buttons into a dialog and use the dialog
as a wrapper, so that the spoiler embed acts like other embed - without
inner tappable widget child, works as expected and no data lose.