fix(*): temporary fallback to raw code on spoiler and hide embeds
Spoiler and hide embeds are causing extream junk on Android, the root
cause seems live deep in flutter_quill or flutter itself, maybe because
of the page pushing/popping or the nested editor makes some Android
native components confusing on the IME state.

This issue was introduced ever since those two tags are landed. Now
sometimes it does not cause junk but in the background log same
messages still here so it's hard to figure out which user action
directly caused it.

Though a new implementation of direct access on these two tags is on the
way, process still lives in early stage and the junk it caused worth
a temporary fix.

This commit made the functionality fallback to raw code mode on Android.

ref: https://github.com/realth000/tsdm_client/issues/134