fix: fix paragraph attributes not finished on last paragraph
For the following example BBCode:

```
[code]
xxx
[/code]
```

the parsed quill delta as a '\n' as the last operation which also have a
code-block attribute. It's not added to meet "always ended with '\n'"
purpose but to finish the code block paragraph attribute.

In previous versions, when `insertBBCode`, the last paragraph would be
removed if it only has a '\n', considered as the meet of "always ended
with '\n'", we shall not that if the paragraph has paragraph attributes:
in this condition, '\n' is to finish the paragraph attribute.

This commit added another condition to skip removing the last operation
if paragraph has attributes.